refactor: 重构docxBlockPatch工具并统一术语
- 移动docxBlockPatch到独立文件block/patch.ts
- 统一工具处理逻辑使用handler方法
- 统一术语命名：代码块→代码，高亮块→高亮
